refactor: move custom layers API requests into the feature folder
